+++ src/main/resources/main_application.properties
@@ -1,2 +1,2 @@
-notForkedRepos.url=https://api.github.com/search/repositories?q=user:Tincoquesse fork:false
-branchesFromRepo.url=https://api.github.com/repos/Tincoquesse/Alledrogo_Frontend/branches
+notForkedRepos.url=http://localhost:8090/search/repositories?q=user:Tincoquesse fork:false
+branchesFromRepo.url=http://localhost:8090/repos/Tincoquesse/Alledrogo_Frontend/branches
