+++ src/test/java/com/atipera/recruitment_task/api/RestSpringbootControllerTest.java
@@ -1,6 +1,5 @@
 package com.atipera.recruitment_task.api;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
@@ -18,8 +17,6 @@ class RestSpringbootControllerTest {
     @Autowired
     private MockMvc mockMvc;
 
-    @Autowired
-    private ObjectMapper objectMapper;
 
     @Test
     public void shouldGetRepos() throws Exception {
