--- src/test/resources/__files/json/branches_response_success.json
@@ -1,1 +1,0 @@
-[{"repoName":"CNC_APP","ownerLogin":"Tincoquesse","branches":[{"branchName":"master","lastCommitSha":"bb06a5084d702c64c271cae92385a9c560aed946"}]},{"repoName":"Kalkulator","ownerLogin":"Tincoquesse","branches":[{"branchName":"master","lastCommitSha":"72829d0ee5702df99458ffd89f498bcc7a7e9104"}]},{"repoName":"Space-web-app","ownerLogin":"Tincoquesse","branches":[{"branchName":"master","lastCommitSha":"70ea155281c50104569053fa3190472affdea284"}]},{"repoName":"spring_fsd_project","ownerLogin":"Tincoquesse","branches":[{"branchName":"Profile","lastCommitSha":"a8485636fd3ff100d9b764c03bc781a0a8fb343f"},{"branchName":"master","lastCommitSha":"0d3d49f77f8b13ac0d169a12740c0f9e62da8842"},{"branchName":"qualifier-branch","lastCommitSha":"519207d3bd6b7da6a0e4abcd87f0b8ff78b17b7e"},{"branchName":"side_branch","lastCommitSha":"ab6ade47c7c344ba74385d33cb6890ef1da69023"},{"branchName":"singleton_prototype_test","lastCommitSha":"c1ae2c5569cd34e50a16679d2c331fb7af5e4824"},{"branchName":"singleton","lastCommitSha":"e02b9947f0b5578c19e176c3bb7d929f9d9d3410"},{"branchName":"wired_branch","lastCommitSha":"c1ae2c5569cd34e50a16679d2c331fb7af5e4824"}]},{"repoName":"StarWars_site","ownerLogin":"Tincoquesse","branches":[{"branchName":"master","lastCommitSha":"e4c879c0b60ccd0fdca0ba47283c2ee15d21abd7"}]},{"repoName":"Recruitment_task","ownerLogin":"Tincoquesse","branches":[{"branchName":"master","lastCommitSha":"c1d3e18e4ebac8e4e3cd7bff9b8eccf704e6c6d0"}]},{"repoName":"Alledrogo_Spring_Backend","ownerLogin":"Tincoquesse","branches":[{"branchName":"feature/order_finalize","lastCommitSha":"e167db5e5c53d93ade7b8cc1acaf9badc4da10ab"},{"branchName":"master","lastCommitSha":"14c1f46e2c5049a61d57bbfb0bb3c7d4df2eb624"}]},{"repoName":"inqoo_angular_basisc","ownerLogin":"Tincoquesse","branches":[{"branchName":"master","lastCommitSha":"8903591f249f0cda03edcde67baa09775979cb27"}]},{"repoName":"Alledrogo_Frontend","ownerLogin":"Tincoquesse","branches":[{"branchName":"feature/order_form","lastCommitSha":"4b5ce83148901b7184068e6cdd91874528b5e0c9"},{"branchName":"feature/searchbar","lastCommitSha":"05aa54d26058a99fe76e182fe820b20c641accce"},{"branchName":"master","lastCommitSha":"1d0f6ada7421132721d9b22cdaac079593bf674f"}]},{"repoName":"SchoolBudget","ownerLogin":"Tincoquesse","branches":[{"branchName":"master","lastCommitSha":"c8954b551da21f656bca5215c1355c99cf69a7ce"}]},{"repoName":"CLI_Banking_System","ownerLogin":"Tincoquesse","branches":[{"branchName":"master","lastCommitSha":"b80abb2df0142207c0b7b1083172971e30b3798d"}]}]
