+++ src/test/resources/branches_response_success.json
@@ -1,0 +1,1 @@
+[{"name":"feature/order_form","commit":{"sha":"4b5ce83148901b7184068e6cdd91874528b5e0c9","url":"https://api.github.com/repos/Tincoquesse/Alledrogo_Frontend/commits/4b5ce83148901b7184068e6cdd91874528b5e0c9"},"protected":false},{"name":"feature/searchbar","commit":{"sha":"05aa54d26058a99fe76e182fe820b20c641accce","url":"https://api.github.com/repos/Tincoquesse/Alledrogo_Frontend/commits/05aa54d26058a99fe76e182fe820b20c641accce"},"protected":false},{"name":"master","commit":{"sha":"1d0f6ada7421132721d9b22cdaac079593bf674f","url":"https://api.github.com/repos/Tincoquesse/Alledrogo_Frontend/commits/1d0f6ada7421132721d9b22cdaac079593bf674f"},"protected":true}]
